% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_demo.r
\name{addNums}
\alias{addNums}
\title{Add 2 numbers!}
\usage{
addNums(nRow, nCol, addSample = TRUE, ...)

addNums(nRow, nCol, addSample = TRUE, ...)
}
\arguments{
\item{x:}{Float/Int}

\item{y:}{Float/Int}

\item{nRow:}{Int}

\item{nCol:}{Int}

\item{addSample:}{Bool - Add a "Sample" column, numbered 1:nRow}
}
\value{
Summed numbers

Data.Frame
}
\description{
Add 2 numbers!

Create random table!
}
