% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_demo.r
\name{calcPCAEllipse}
\alias{calcPCAEllipse}
\title{Calculate the ellipse for a confidence interval on the PCA projections}
\usage{
calcPCAEllipse(pcaProjections, confidence = 95, ...)
}
\arguments{
\item{pcaProjections:}{Scores from a prcomp() fit}

\item{confidence:}{Hotellings confidence (usually 95\%)}
}
\value{
2D Ellipse boundaries for plotting the confidence interval
}
\description{
Calculate the ellipse for a confidence interval on the PCA projections
}
