% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_demo.r
\name{createRandomTbl}
\alias{createRandomTbl}
\title{Create random table!}
\usage{
createRandomTbl(nRow, nCol, addSample = TRUE, ...)
}
\arguments{
\item{nRow:}{Int}

\item{nCol:}{Int}

\item{addSample:}{Bool - Add a "Sample" column, numbered 1:nRow}
}
\value{
Data.Frame
}
\description{
Create random table!
}
