% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_demo.r
\name{fitPCA}
\alias{fitPCA}
\title{Fit PCA on data.frame}
\usage{
fitPCA(data, nonDatCols = c("Sample"), returnFit = FALSE)
}
\arguments{
\item{data:}{Individual MZ data aggregated into NxM}

\item{nonDatCols:}{List - Non-data-containing columns}

\item{returnFit:}{List - Return only the fit object}
}
\value{
PCA scores
}
\description{
Note: columns are removed which have 0 variance
}
