\name{foo}
\alias{foo}
\title{The foo function}
\usage{
  foo(a, b)
}
\arguments{
  \item{a}{one argument}

  \item{b}{the other one}
}
\value{
  \code{NULL}
}
\description{
  This is the foo function.
}
\section{Special Section}{
  This is a customized \bold{section}.
  
  Another paragraph with a URL \url{https://github.com/yihui/Rd2roxygen}.
}
\examples{
foo(1, 2)
}
\author{
  Yihui Xie <\url{http://yihui.org}>
}

